pyproject.toml
[project]
name = "midi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anywidget[dev]>=0.9.21",
    "langchain>=1.2.2",
    "langchain-anthropic>=1.3.1",
    "langchain-openai>=1.1.7",
    "marimo[recommended]>=0.18.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.10",
    "wandb>=0.23.1",
    "weave>=0.52.22",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
